requirements.txt
Flask
requests
spotipy
openai
python-dotenv